BGDIINF_SB-1723: Moved TestHttp500 in its own module
This class is only for unittest purpose therefore move it into its own
module.
